KVM: PPC: Used long unsigned for index in H_PUT_TCE_INDIRECT
This changes some variables to long type in order that
(i << tbl->it_page_shift) does not overflow and evaluate to a negative
value.

+++ arch/powerpc/kvm/book3s_64_vio.c
@@ -399,7 +399,7 @@ static long kvmppc_h_put_tce_indirect_iommu(struct kvm_vcpu *vcpu,
 		struct iommu_table *tbl, unsigned long ioba,
 		u64 __user *tces, unsigned long npages)
 {
-	int i, ret;
+	unsigned long i, ret;
 	const unsigned long entry = ioba >> tbl->it_page_shift;
 	unsigned long tce, gpa;
